+++ jniLibs/MobileRT/src/main/cpp/MobileRT/Accelerators/BVH2.hpp
@@ -21,7 +21,8 @@ namespace MobileRT {
       ::std::vector<::std::vector<MobileRT::Primitive<T>>> primitives_ {};
 
     private:
-      AABB build(::std::vector<MobileRT::Primitive<T>>&& primitives, const unsigned depth, const uint64_t currentNodeId) noexcept {
+      AABB build(::std::vector<MobileRT::Primitive<T>> &&primitives, const unsigned depth,
+                 const uint32_t currentNodeId) noexcept {
         if (primitives.empty()) {
           if (depth == 0) {
             boxes_.emplace_back(AABB{});
@@ -42,7 +43,7 @@ namespace MobileRT {
             std::numeric_limits<float>::lowest(),
             std::numeric_limits<float>::lowest(),
             std::numeric_limits<float>::lowest()}};
-        for (uint64_t i {0}; i < primitives.size(); i++) {
+          for (uint32_t i{0}; i < primitives.size(); i++) {
           const AABB new_box {primitives.at(i).getAABB()};
           current_box = surroundingBox(new_box, current_box);
         }
@@ -68,7 +69,7 @@ namespace MobileRT {
             break;
         }
 
-        int64_t divide {static_cast<int64_t>(primitives.size()) / 2};
+          int32_t divide{static_cast<int32_t>(primitives.size()) / 2};
         /*uint64_t divide {0};
         {
           std::vector<float> left_area {boxes.at(0).getSurfaceArea()};
@@ -98,9 +99,9 @@ namespace MobileRT {
         }*/
         divide = primitives.size() % 2 == 0? divide : divide + 1;
 
-        const uint64_t aux {static_cast<uint64_t>(1 << depth)};
-        const uint64_t left {currentNodeId * 2 + 1};
-        const uint64_t right {left + 1};
+          const uint32_t aux{static_cast<uint32_t>(1 << depth)};
+          const uint32_t left{currentNodeId * 2 + 1};
+          const uint32_t right{left + 1};
 
         //LOG("depth = ", depth, ", currentNodeId = ", currentNodeId,", size = ", primitives.size(), ", divide = ", divide);
         //LOG("leftChild = ", left, ", rightChild = ", right);
@@ -129,7 +130,8 @@ namespace MobileRT {
     public:
       explicit BVH2 () noexcept = default;
 
-      explicit BVH2<T> (::std::vector<MobileRT::Primitive<T>>&& primitives, const unsigned depth = 0, const uint64_t currentId = 0) noexcept {
+      explicit BVH2<T>(::std::vector<MobileRT::Primitive<T>> &&primitives, const unsigned depth = 0,
+                       const uint32_t currentId = 0) noexcept {
 
         if (primitives.empty()) {
           if (depth == 0) {
@@ -137,12 +139,12 @@ namespace MobileRT {
           }
           return;
         }
-        uint64_t numberPrimitives {primitives.size()};
-        uint64_t maxDepth {0};
-        uint64_t maxNodes {1};
+          uint32_t numberPrimitives{static_cast<uint32_t>(primitives.size())};
+          uint32_t maxDepth{0};
+          uint32_t maxNodes{1};
         while (maxNodes * 2 < numberPrimitives) {
           maxDepth++;
-          maxNodes = 1 << maxDepth;
+            maxNodes = 1u << maxDepth;
         }
         while (maxDepth > 0) {
           maxDepth--;
@@ -164,12 +166,13 @@ namespace MobileRT {
 
       BVH2 &operator= (BVH2 &&bVH) noexcept = default;
 
-      bool trace (::MobileRT::Intersection *const intersection, const ::MobileRT::Ray &ray, const unsigned depth = 0, const uint64_t currentNodeId = 0) noexcept {
+      bool trace(::MobileRT::Intersection *const intersection, const ::MobileRT::Ray &ray,
+                 const unsigned depth = 0, const uint32_t currentNodeId = 0) noexcept {
         if (boxes_.at(currentNodeId).intersect(ray)) {
           //node at the bottom of tree - no childs
-          const uint64_t aux {static_cast<uint64_t>(1 << depth)};
+            const uint32_t aux{static_cast<uint32_t>(1 << depth)};
           if (depth == numberDepth_) {
-            const uint64_t primitiveId {(currentNodeId - (aux - 1))};
+              const uint32_t primitiveId{(currentNodeId - (aux - 1))};
             bool res {false};
             for (MobileRT::Primitive<T> &primitive : primitives_.at(primitiveId)) {
               res |= primitive.intersect(intersection, ray);
@@ -177,7 +180,7 @@ namespace MobileRT {
             return res;
           }
 
-          const uint64_t left {currentNodeId * 2 + 1};
+            const uint32_t left{currentNodeId * 2 + 1};
           const bool hit_left {trace(intersection, ray, depth + 1, left)};
           const bool hit_right {trace(intersection, ray, depth + 1, left + 1)};
           return hit_left || hit_right;
@@ -185,12 +188,13 @@ namespace MobileRT {
         return false;
       }
 
-      bool shadowTrace (::MobileRT::Intersection *const intersection, const ::MobileRT::Ray &ray, const unsigned depth = 0, const uint64_t currentNodeId = 0) noexcept {
+      bool shadowTrace(::MobileRT::Intersection *const intersection, const ::MobileRT::Ray &ray,
+                       const unsigned depth = 0, const uint32_t currentNodeId = 0) noexcept {
         if (boxes_.at(currentNodeId).intersect(ray)) {
           //node at the bottom of tree - no childs
-          const uint64_t aux {static_cast<uint64_t>(1 << depth)};
+            const uint32_t aux{static_cast<uint32_t>(1 << depth)};
           if (depth == numberDepth_) {
-            const uint64_t primitiveId {(currentNodeId - (aux - 1))};
+              const uint32_t primitiveId{(currentNodeId - (aux - 1))};
             for (MobileRT::Primitive<T> &primitive : primitives_.at(primitiveId)) {
               if (primitive.intersect(intersection, ray)) {
                 return true;
@@ -199,7 +203,7 @@ namespace MobileRT {
             return false;
           }
 
-          const uint64_t left {currentNodeId * 2 + 1};
+            const uint32_t left{currentNodeId * 2 + 1};
           const bool hit_left {shadowTrace(intersection, ray, depth + 1, left)};
           const bool hit_right {shadowTrace(intersection, ray, depth + 1, left + 1)};
           return hit_left || hit_right;
